+++ contracts/DAITokenWrapper.sol
@@ -3,6 +3,7 @@ pragma solidity 0.7.6;
 
 
 import "./interfaces/ITokenWrapper.sol";
+import "./interfaces/IDAI.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 /**
@@ -92,24 +93,4 @@ contract DAITokenWrapper is
     {
         return daiTokenAddress;
     }
-}
-
-/**
- * @title IDAI
- * @notice Interface for the purpose of calling the permit function on the deployed DAI token
- */
-interface IDAI {
- 
-    function name() external pure returns (string memory);
-    
-    function permit(address holder, address spender, uint256 nonce, uint256 expiry,
-                    bool allowed, uint8 v, bytes32 r, bytes32 s) external;
-
-    function transferFrom(
-        address from,
-        address to,
-        uint256 value
-    ) external returns (bool);
-
-    function nonces(address owner) external view returns (uint256);
 }
